requirementsWindows.txt
# knihovny potřebné pro běh programu ve Windows
opencv-python
numpy
paho-mqtt
keyboard
pysimplegui
queuelib